# Optimize DependencyProperty lookup/registration performance, remove allocs