--- a/src/Microsoft.DotNet.Wpf/src/WindowsBase/System/Windows/DependencyProperty.cs
+++ b/src/Microsoft.DotNet.Wpf/src/WindowsBase/System/Windows/DependencyProperty.cs
@@ -1189,7 +1189,7 @@ namespace System.Windows
         private static int GlobalIndexCount;
 
         // Global, cross-object synchronization
-        internal static readonly object Synchronized = new object();
+        internal static readonly Lock Synchronized = new();
 
         // Nullable Type
         private static Type NullableType = typeof(Nullable<>);